# openapi.json
{
  "components": {
    "schemas": {
      "CaptureIn": {
        "description": "Body of `POST /v1/invoices/{invoice_id}/captures`.",
        "properties": {
          "amount_cents": {
            "exclusiveMinimum": 0.0,
            "title": "Amount Cents",
            "type": "integer"
          }
        },
        "required": [
          "amount_cents"
        ],
        "title": "CaptureIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InvoiceIn": {
        "description": "Body of `POST /v1/invoices`.",
        "properties": {
          "customer_email": {
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Customer Email",
            "type": "string"
          },
          "discount_percent": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Discount Percent",
            "type": "number"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LineItemIn"
            },
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "region": {
            "title": "Region",
            "type": "string"
          }
        },
        "required": [
          "customer_email",
          "region",
          "items"
        ],
        "title": "InvoiceIn",
        "type": "object"
      },
      "InvoiceOut": {
        "description": "An invoice with its priced totals.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "customer_email": {
            "title": "Customer Email",
            "type": "string"
          },
          "discount_cents": {
            "title": "Discount Cents",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "subtotal_cents": {
            "title": "Subtotal Cents",
            "type": "integer"
          },
          "tax_cents": {
            "title": "Tax Cents",
            "type": "integer"
          },
          "total_cents": {
            "title": "Total Cents",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "customer_email",
          "region",
          "status",
          "subtotal_cents",
          "discount_cents",
          "tax_cents",
          "total_cents",
          "created_at"
        ],
        "title": "InvoiceOut",
        "type": "object"
      },
      "LineItemIn": {
        "description": "One billable line on a new invoice.",
        "properties": {
          "description": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "quantity": {
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          },
          "unit_amount_cents": {
            "minimum": 0.0,
            "title": "Unit Amount Cents",
            "type": "integer"
          }
        },
        "required": [
          "description",
          "quantity",
          "unit_amount_cents"
        ],
        "title": "LineItemIn",
        "type": "object"
      },
      "PaymentOut": {
        "description": "A capture or a refund.",
        "properties": {
          "amount_cents": {
            "title": "Amount Cents",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "invoice_id": {
            "title": "Invoice Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "invoice_id",
          "kind",
          "amount_cents",
          "created_at"
        ],
        "title": "PaymentOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Ledgerline",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/health": {
      "get": {
        "description": "Liveness. Does not touch the database.",
        "operationId": "health_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health V1 Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/v1/health/ready": {
      "get": {
        "description": "Readiness. Runs a trivial query against the database.",
        "operationId": "ready_v1_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Ready V1 Health Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready",
        "tags": [
          "health"
        ]
      }
    },
    "/v1/invoices": {
      "post": {
        "description": "Price and store a new invoice.",
        "operationId": "create_invoice_v1_invoices_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-role",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-User-Role",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Invoice",
        "tags": [
          "invoices"
        ]
      }
    },
    "/v1/invoices/list": {
      "get": {
        "description": "Every invoice, newest first.",
        "operationId": "list_invoices_v1_invoices_list_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-role",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-User-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvoiceOut"
                  },
                  "title": "Response List Invoices V1 Invoices List Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Invoices",
        "tags": [
          "invoices"
        ]
      }
    },
    "/v1/invoices/{invoice_id}": {
      "get": {
        "description": "One invoice by id.",
        "operationId": "get_invoice_v1_invoices__invoice_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "title": "Invoice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-role",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-User-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Invoice",
        "tags": [
          "invoices"
        ]
      }
    },
    "/v1/invoices/{invoice_id}/captures": {
      "post": {
        "description": "Capture money against an open invoice.",
        "operationId": "capture_payment_v1_invoices__invoice_id__captures_post",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "title": "Invoice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-role",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-User-Role",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Capture Payment",
        "tags": [
          "payments"
        ]
      }
    },
    "/v1/invoices/{invoice_id}/payments": {
      "get": {
        "description": "Captures and refunds recorded for an invoice.",
        "operationId": "list_payments_v1_invoices__invoice_id__payments_get",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "title": "Invoice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-role",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-User-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PaymentOut"
                  },
                  "title": "Response List Payments V1 Invoices  Invoice Id  Payments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Payments",
        "tags": [
          "payments"
        ]
      }
    }
  }
}